@@ -20,18 +20,18 @@
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "^13.2.5",
-        "@angular/cdk": "^13.2.5",
-        "@angular/common": "^13.2.5",
-        "@angular/compiler": "^13.2.5",
-        "@angular/core": "^13.2.5",
-        "@angular/flex-layout": "^13.0.0-beta.38",
-        "@angular/forms": "^13.2.5",
-        "@angular/localize": "^13.2.5",
-        "@angular/material": "^13.2.5",
-        "@angular/platform-browser": "^13.2.5",
-        "@angular/platform-browser-dynamic": "^13.2.5",
-        "@angular/router": "^13.2.5",
+        "@angular/animations": "^14.1.1",
+        "@angular/cdk": "^14.1.1",
+        "@angular/common": "^14.1.1",
+        "@angular/compiler": "^14.1.1",
+        "@angular/core": "^14.1.1",
+        "@angular/flex-layout": "^14.0.0-beta.40",
+        "@angular/forms": "^14.1.1",
+        "@angular/localize": "^14.1.1",
+        "@angular/material": "^14.1.1",
+        "@angular/platform-browser": "^14.1.1",
+        "@angular/platform-browser-dynamic": "^14.1.1",
+        "@angular/router": "^14.1.1",
         "@apollo/client": "^3.5.10",
         "@graphql-tools/mock": "^8.5.2",
         "apollo-angular": "^3.0.0",
@@ -56,20 +56,20 @@
         "zone.js": "~0.11.4"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "^13.2.5",
-        "@angular-eslint/builder": "13.1.0",
-        "@angular-eslint/eslint-plugin": "13.1.0",
-        "@angular-eslint/eslint-plugin-template": "13.1.0",
-        "@angular-eslint/schematics": "13.1.0",
-        "@angular-eslint/template-parser": "13.1.0",
-        "@angular/cli": "^13.2.5",
-        "@angular/compiler-cli": "^13.2.5",
-        "@angular/language-service": "^13.2.5",
+        "@angular-devkit/build-angular": "^14.1.1",
+        "@angular-eslint/builder": "14.0.2",
+        "@angular-eslint/eslint-plugin": "14.0.2",
+        "@angular-eslint/eslint-plugin-template": "14.0.2",
+        "@angular-eslint/schematics": "14.0.2",
+        "@angular-eslint/template-parser": "14.0.2",
+        "@angular/cli": "^14.1.1",
+        "@angular/compiler-cli": "^14.1.1",
+        "@angular/language-service": "^14.1.1",
         "@playwright/test": "^1.19.2",
-        "@types/jasmine": "~3.10.3",
+        "@types/jasmine": "~4.0.3",
         "@types/jasminewd2": "^2.0.10",
         "@types/lodash-es": "^4.17.6",
-        "@types/node": "^14.14.28",
+        "@types/node": "^16.0",
         "@types/prosemirror-commands": "^1.0.4",
         "@types/prosemirror-dropcursor": "^1.0.3",
         "@types/prosemirror-gapcursor": "^1.0.4",
@@ -82,19 +82,19 @@
         "@types/prosemirror-schema-list": "^1.0.3",
         "@types/prosemirror-state": "^1.2.7",
         "@types/prosemirror-view": "^1.23.1",
-        "@typescript-eslint/eslint-plugin": "5.13.0",
-        "@typescript-eslint/parser": "5.13.0",
-        "eslint": "^8.2.0",
-        "jasmine-core": "~4.0.1",
+        "@typescript-eslint/eslint-plugin": "^5.29.0",
+        "@typescript-eslint/parser": "^5.29.0",
+        "eslint": "^8.18.0",
+        "jasmine-core": "~4.3.0",
         "jasmine-spec-reporter": "~7.0.0",
-        "karma": "~6.3.17",
+        "karma": "~6.4.0",
         "karma-chrome-launcher": "~3.1.0",
         "karma-coverage": "^2.2.0",
-        "karma-jasmine": "~5.0.0",
-        "karma-jasmine-html-reporter": "^1.7.0",
-        "ng-packagr": "^13.2.1",
-        "prettier": "2.5.1",
+        "karma-jasmine": "~5.1.0",
+        "karma-jasmine-html-reporter": "^2.0.0",
+        "ng-packagr": "^14.1.0",
+        "prettier": "2.7.1",
         "ts-node": "^10.6.0",
-        "typescript": "~4.5.5"
+        "typescript": "~4.7.4"
     }
 }
