@@ -14,10 +14,10 @@
         "tslib": "^2.3.1"
     },
     "peerDependencies": {
-        "@angular/common": "^13.2.0",
-        "@angular/core": "^13.2.0",
-        "@angular/material": "^13.2.0",
-        "@angular/localize": "^13.2.0",
+        "@angular/common": "^14.1.0",
+        "@angular/core": "^14.1.0",
+        "@angular/material": "^14.1.0",
+        "@angular/localize": "^14.1.0",
         "apollo-angular": "^2.6.0",
         "lodash-es": "^4.17.21",
         "rxjs": "^7.4.0"
