@@ -1,6 +1,6 @@
 import {Component, Inject} from '@angular/core';
 import {MAT_DIALOG_DATA, MatDialogRef} from '@angular/material/dialog';
-import {FormControl, FormGroup, Validators} from '@angular/forms';
+import {UntypedFormControl, UntypedFormGroup, Validators} from '@angular/forms';
 import {ifValid} from '@ecodev/natural';
 
 export interface IdDialogData {
@@ -20,8 +20,8 @@ export interface IdDialogData {
     styleUrls: ['./id-dialog.component.scss'],
 })
 export class IdDialogComponent {
-    public readonly idControl = new FormControl('', Validators.pattern(/(^(-?[_a-zA-Z]+[_a-zA-Z0-9-]*)+)/));
-    public readonly form = new FormGroup({
+    public readonly idControl = new UntypedFormControl('', Validators.pattern(/(^(-?[_a-zA-Z]+[_a-zA-Z0-9-]*)+)/));
+    public readonly form = new UntypedFormGroup({
         id: this.idControl,
     });
 
