@@ -3,7 +3,7 @@ import {BehaviorSubject, merge, Observable} from 'rxjs';
 import {FilterGroupConditionField} from '../search/classes/graphql-doctrine.types';
 import {NATURAL_DROPDOWN_DATA, NaturalDropdownData} from '../search/dropdown-container/dropdown.service';
 import {DropdownComponent} from '../search/types/dropdown-component';
-import {FormControl, FormGroup, ValidatorFn, Validators} from '@angular/forms';
+import {UntypedFormControl, UntypedFormGroup, ValidatorFn, Validators} from '@angular/forms';
 import {PossibleDiscreteOperator, possibleDiscreteOperators} from './types';
 import {startWith} from 'rxjs/operators';
 
@@ -14,9 +14,9 @@ export abstract class AbstractAssociationSelectComponent<C> implements DropdownC
 
     public requireValueCtrl = false;
     public readonly operators = possibleDiscreteOperators;
-    public readonly operatorCtrl: FormControl = new FormControl('is');
-    public readonly valueCtrl: FormControl = new FormControl();
-    public readonly form = new FormGroup({
+    public readonly operatorCtrl: UntypedFormControl = new UntypedFormControl('is');
+    public readonly valueCtrl: UntypedFormControl = new UntypedFormControl();
+    public readonly form = new UntypedFormGroup({
         operator: this.operatorCtrl,
         value: this.valueCtrl,
     });
