@@ -1,5 +1,5 @@
 import {Component, Inject} from '@angular/core';
-import {FormControl, FormGroupDirective, NgForm, Validators} from '@angular/forms';
+import {UntypedFormControl, FormGroupDirective, NgForm, Validators} from '@angular/forms';
 import {ErrorStateMatcher} from '@angular/material/core';
 import {BehaviorSubject} from 'rxjs';
 import {FilterGroupConditionField} from '../../search/classes/graphql-doctrine.types';
@@ -8,7 +8,7 @@ import {NATURAL_DROPDOWN_DATA, NaturalDropdownData} from '../../search/dropdown-
 import {DropdownComponent} from '../../search/types/dropdown-component';
 
 export class InvalidWithValueStateMatcher implements ErrorStateMatcher {
-    public isErrorState(control: FormControl | null, form: FormGroupDirective | NgForm | null): boolean {
+    public isErrorState(control: UntypedFormControl | null, form: FormGroupDirective | NgForm | null): boolean {
         return control && control.invalid && control.value;
     }
 }
@@ -19,7 +19,7 @@ export class InvalidWithValueStateMatcher implements ErrorStateMatcher {
 })
 export class TypeTextComponent implements DropdownComponent {
     public renderedValue = new BehaviorSubject<string>('');
-    public formCtrl: FormControl = new FormControl();
+    public formCtrl: UntypedFormControl = new UntypedFormControl();
     public matcher = new InvalidWithValueStateMatcher();
 
     public constructor(
